Refactor: Rename package from "com.sriky.jokeproviderlib" to "com.sriky.jokeprovider"
